Add files via upload
The prevalence of diabetic retinopathy is rising worldwide, requiring effective early detection and treatment strategies. Our research attempts to improve the accuracy of screening and diagnosis by utilizing machine learning techniques. We enhance accessibility and results in the management of diabetic retinopathy by compiling large datasets and highlighting the importance of efficient screening instruments.
The goal of this study is to discuss the difficulties in effectively diagnosing and treating diabetic retinopathy, one of the main effects of diabetes.